@@ -4,7 +4,7 @@
 
 <script>
   import { mapState } from 'vuex'
-  import firebase from 'firebase'
+  import { getAuth, GoogleAuthProvider, EmailAuthProvider } from 'firebase/auth'
   import firebaseui from 'firebaseui'
   import firebaseApp from '../firebase-app'
 
@@ -22,11 +22,11 @@
     credentialHelper: firebaseui.auth.CredentialHelper.NONE,
     signInOptions: [
       {
-        provider: firebase.auth.GoogleAuthProvider.PROVIDER_ID,
+        provider: GoogleAuthProvider.PROVIDER_ID,
         scopes: [ 'https://www.googleapis.com/auth/plus.login' ]
       },
       {
-        provider: firebase.auth.EmailAuthProvider.PROVIDER_ID,
+        provider: EmailAuthProvider.PROVIDER_ID,
         requireDisplayName: true
       }
     ]
@@ -34,13 +34,14 @@
 
   export default {
     mounted () {
-      ui = ui || new firebaseui.auth.AuthUI(firebaseApp.auth())
+      const auth = getAuth(firebaseApp)
+      ui = ui || new firebaseui.auth.AuthUI(auth)
       ui.start('#firebaseui-auth', uiConfig)
     },
     computed: {
       ...mapState([ 'user' ])
     },
-    destroyed () {
+    unmounted () {
       ui.reset()
     },
     watch: {
@@ -58,25 +59,25 @@
 </script>
 
 <style lang="sass">
-  @import '~firebaseui/dist/firebaseui.css'
+@import 'firebaseui/dist/firebaseui.css'
 
-  ul.firebaseui-idp-list
-    margin: 0
-    list-style-type: none
+ul.firebaseui-idp-list
+  margin: 0
+  list-style-type: none
 
-  .firebaseui-card-content
-    padding: 0 5px
+.firebaseui-card-content
+  padding: 0 5px
 
-  .firebaseui-container
-    max-width: 768px
+.firebaseui-container
+  max-width: 768px
 
-  .firebaseui-idp-button
-    font-weight: 100
-    max-width: 768px
-    line-height: 36px
-    span
-      text-align: center
+.firebaseui-idp-button
+  font-weight: 100
+  max-width: 768px
+  line-height: 36px
+  span
+    text-align: center
 
-  .firebaseui-idp-google > .firebaseui-idp-text
-    color: #363636
+.firebaseui-idp-google > .firebaseui-idp-text
+  color: #363636
 </style>
