@@ -1,41 +1,39 @@
-import Vue from 'vue'
-import Router from 'vue-router'
+import { createRouter, createWebHistory } from 'vue-router'
 
-import UserUpdate from '@/components/UserUpdate'
-import Auth from '@/components/Auth'
+import UserUpdate from '@/components/UserUpdate.vue'
+import Auth from '@/components/Auth.vue'
 
-import About from '@/components/About'
+import About from '@/components/About.vue'
 // import Overview from '@/components/public/About'
-import Download from '@/components/public/Download'
-import ArtworkList from '@/components/public/ArtworkList'
-import ArtworkDetail from '@/components/public/Artwork'
-import ArtistDetail from '@/components/public/Artist'
-import ShowDetail from '@/components/public/Show'
-import PlaceDetail from '@/components/public/Place'
-import PlayerClient from '@/components/public/PlayerClient'
+import Download from '@/components/public/Download.vue'
+import ArtworkList from '@/components/public/ArtworkList.vue'
+import ArtworkDetail from '@/components/public/Artwork.vue'
+import ArtistDetail from '@/components/public/Artist.vue'
+import ShowDetail from '@/components/public/Show.vue'
+import PlaceDetail from '@/components/public/Place.vue'
+import PlayerClient from '@/components/public/PlayerClient.vue'
 
-import AccountNew from '@/components/account/AccountNew'
-import AccountDetail from '@/components/account/AccountDetail'
-import AccountInvitationList from '@/components/account/InvitationList'
+import AccountNew from '@/components/account/AccountNew.vue'
+import AccountDetail from '@/components/account/AccountDetail.vue'
+import AccountInvitationList from '@/components/account/InvitationList.vue'
 
-import AccountPlayerNew from '@/components/account/PlayerNew'
-import AccountPlayerList from '@/components/account/PlayerList'
-import AccountArtworkIterationDetail from '@/components/account/IterationDetail'
-import AccountArtworkEditor from '@/components/account/ArtworkEditor'
-import AccountArtistList from '@/components/account/ArtistList'
-import AccountArtistDetail from '@/components/account/ArtistDetail'
-import AccountArtistEditor from '@/components/account/ArtistEditor'
-import AccountShowList from '@/components/account/ShowList'
-import AccountShowDetail from '@/components/account/ShowDetail'
-import AccountShowEditor from '@/components/account/ShowEditor'
-import AccountPlaceList from '@/components/account/PlaceList'
-import AccountPlaceDetail from '@/components/account/PlaceDetail'
-import AccountPlaceEditor from '@/components/account/PlaceEditor'
+import AccountPlayerNew from '@/components/account/PlayerNew.vue'
+import AccountPlayerList from '@/components/account/PlayerList.vue'
+import AccountArtworkIterationDetail from '@/components/account/IterationDetail.vue'
+import AccountArtworkEditor from '@/components/account/ArtworkEditor.vue'
+import AccountArtistList from '@/components/account/ArtistList.vue'
+import AccountArtistDetail from '@/components/account/ArtistDetail.vue'
+import AccountArtistEditor from '@/components/account/ArtistEditor.vue'
+import AccountShowList from '@/components/account/ShowList.vue'
+import AccountShowDetail from '@/components/account/ShowDetail.vue'
+import AccountShowEditor from '@/components/account/ShowEditor.vue'
+import AccountPlaceList from '@/components/account/PlaceList.vue'
+import AccountPlaceDetail from '@/components/account/PlaceDetail.vue'
+import AccountPlaceEditor from '@/components/account/PlaceEditor.vue'
 
-Vue.use(Router)
-
-export function createRouter () {
-  return new Router({
+export function createAppRouter () {
+  const router = createRouter({
+    history: createWebHistory(),
     routes: [
       {path: '/', component: About, name: 'home'},
       {path: '/user/update', component: UserUpdate, name: 'user-update'},
@@ -79,4 +77,6 @@ export function createRouter () {
       {path: '/account/place/:id', component: AccountPlaceDetail}
     ]
   })
+
+  return router
 }
