@@ -92,8 +92,8 @@
 <script>
   import { mapActions, mapState } from 'vuex'
   import firebase from '../../firebase-app'
-  import AttachmentPanel from '../elements/AttachmentsPanel'
-  import RemoteControl from '../elements/RemoteControl'
+  import AttachmentPanel from '../elements/AttachmentsPanel.vue'
+  import RemoteControl from '../elements/RemoteControl.vue'
   import { log } from '../../helper'
   import { PlayerArtwork } from '../../models/PlayerArtwork'
   import { Artwork } from '../../models/ArtworkData'
