@@ -1,4 +1,3 @@
-import Vue from 'vue'
 import { firebaseMutations } from 'vuexfire'
 
 export default {
@@ -37,7 +36,7 @@ export default {
   SET_ITEMS: (state, {items}) => {
     items.forEach(item => {
       if (item) {
-        Vue.set(state.items, item.__key, item)
+        state.items[item.__key] = item
       }
     })
   },
